cleaned up a few parts of even more guillibles travels and personal hygiene
